+++ src/Omega_h_file.hpp
@@ -135,7 +135,7 @@ INST_DECL(I64)
 INST_DECL(Real)
 #undef INST_DECL
 // for VTK compression headers
-extern template void swap_if_needed(std::size_t& val, bool is_little_endian);
+extern template void swap_if_needed(std::uint64_t& val, bool is_little_endian);
 }  // namespace binary
 
 inline std::string to_string(I32 x) {
